del msg fix with roles

+++ server/src/req_handlers/mx_handle_delete_chat_msg.c
@@ -74,9 +74,63 @@ static inline void notify_all_members(client_t* client,
 }
 
 
+static inline bool user_has_permission_to_del_message(const int chat_id, const int message_id, const int user_id){
+/*
+SELECT rm1.permissions AS user_permission, m.from_id, rm2.permissions AS from_permission
+FROM message AS m
+JOIN room_member AS rm1 ON rm1.client_id = <user_id> AND rm1.room_id = <room_id>
+JOIN room_member AS rm2 ON rm2.client_id = m.from_id AND rm2.room_id = <room_id>
+WHERE m.id = <message_id>;
+
+
+SELECT rm1.permissions AS user_permission, m.from_id, rm2.permissions AS from_permission FROM message AS m JOIN room_member AS rm1 ON rm1.client_id = 1 AND rm1.room_id = 2 JOIN room_member AS rm2 ON rm2.client_id = m.from_id AND rm2.room_id = 2 WHERE m.id = 3
+
+SELECT rm1.permissions AS user_permission, m.from_id, rm2.permissions AS from_permission FROM message AS m JOIN room_member AS rm1 ON rm1.client_id = 1 AND rm1.room_id = 1 JOIN room_member AS rm2 ON rm2.client_id = m.from_id AND rm2.room_id = 1 WHERE m.id = 1
+
+	*/
+
+	bool result = false;
+
+	sqlite3* db = mx_open_db();
+
+	pthread_mutex_lock(&db_mutex);
+
+	char* sql_req = NULL;
+	asprintf(&sql_req, "SELECT rm1.permissions AS user_permission, m.from_id, rm2.permissions AS from_permission"
+		               " FROM message AS m JOIN room_member AS rm1 ON rm1.client_id = '%d' AND rm1.room_id = '%d'"
+		               " JOIN room_member AS rm2 ON rm2.client_id = m.from_id AND rm2.room_id = '%d' WHERE m.id = '%d'",
+		                user_id, chat_id, chat_id, message_id);
+	sqlite3_stmt* stmt;
+
+	if (sqlite3_prepare_v2(db, sql_req, -1, &stmt, 0) != SQLITE_OK) {
+	 	mx_log(SERV_LOG_FILE, LOG_ERROR, sqlite3_errmsg(db));
+        mx_close_db(db);
+        exit(-1);
+    }
+
+
+    while(sqlite3_step(stmt) == SQLITE_ROW) {
+    	const int this_user_permissions = sqlite3_column_int64(stmt, 0);
+    	const int message_from_user_id = sqlite3_column_int64(stmt, 1);
+    	const int message_owner_permissions = sqlite3_column_int64(stmt, 2);
+
+    	if(user_id == message_from_user_id || this_user_permissions > message_owner_permissions){
+    		result = true;
+    	}
+    }
+
+
+    free(sql_req);
+    sqlite3_finalize(stmt);
+
+    pthread_mutex_unlock(&db_mutex);
+    mx_close_db(db);
+
+    return result;
+}
+
 void mx_handle_delete_chat_msg(client_t* client, request_t* req) {
 
-	
 
 	mx_log(SERV_LOG_FILE, LOG_TRACE, "Handling delete message request");
 
@@ -87,6 +141,11 @@ void mx_handle_delete_chat_msg(client_t* client, request_t* req) {
 		return;
 
 	char* sql_req = NULL;
+
+	//TODO: check if user has permission to delete this message
+	if(!user_has_permission_to_del_message(chat_id, msg_id, client->user_id)){
+		return;
+	}
 
 	asprintf(&sql_req, "DELETE FROM 'message' WHERE message.id = '%d'", msg_id);
